feat(files): Give files pointers to parent file trees
