+++ bids-validator/src/types/filetree.ts
@@ -1,7 +1,23 @@
 /**
  * Abstract FileTree for all environments (Deno, Browser, Python)
  */
-import { BIDSFile } from '../types/file.ts'
+export interface BIDSFile {
+  // Filename
+  name: string
+  // Dataset relative path for the file
+  path: string
+  // File size in bytes
+  size: number
+  // BIDS ignore status of the file
+  ignored: boolean
+  // ReadableStream to file raw contents
+  stream: ReadableStream<Uint8Array>
+  // Resolve stream to decoded utf-8 text
+  text: () => Promise<string>
+  // Read a range of bytes
+  readBytes: (size: number, offset?: number) => Promise<Uint8Array>
+  parent?: FileTree
+}
 
 export class FileTree {
   // Relative path to this FileTree location
