+++ bids-validator/src/validators/isBidsy.ts
@@ -5,7 +5,7 @@
 // @ts-nocheck
 import { BIDSContext } from '../schema/context.ts'
 import { CheckFunction } from '../../types/check.ts'
-import { BIDSFile } from '../types/file.ts'
+import { BIDSFile } from '../types/filetree.ts'
 import { Schema } from '../types/schema.ts'
 
 export const isBidsyFilename: CheckFunction = (schema, context) => {
